Revert "Feature/basicstests"
